requirements.txt
streamlit~=1.41.1
pdfplumber~=0.11.5
pillow~=11.1.0
PyMuPdf~=1.25.3